now subViews have own folder
